+++ packages/js-toolkit/helpers/createApp.ts
@@ -11,7 +11,7 @@ export type CreateAppOptions = Partial<Features> & {
  * Instantiate and mount the given component on the given root element when the page has been loaded
  * and return a function to use the app instance when it is ready.
  */
-export default function createApp<S extends BaseConstructor<Base>, T extends BaseProps = BaseProps>(
+export function createApp<S extends BaseConstructor<Base>, T extends BaseProps = BaseProps>(
   App: S,
   options: HTMLElement | CreateAppOptions = {},
 ): () => Promise<S & Base<T>> {