+++ packages/js-toolkit/utils/math/clamp.ts
@@ -6,7 +6,7 @@
  * @param {number} max
  * @returns {number}
  */
-export default function clamp(value: number, min: number, max: number) {
+export function clamp(value: number, min: number, max: number) {
   /* eslint-disable no-nested-ternary */
   return min < max
     ? value < min